move consensus switching into separate module
